'use client' vs 'use server'
